rearranged some common ptycho methods to base_class
